+++ com.unity.render-pipelines.universal/Runtime/Passes/PostProcessPass.cs
@@ -64,6 +64,10 @@ namespace UnityEngine.Rendering.Universal.Internal
         // If there's a final post process pass after this pass.
         // If yes, Film Grain and Dithering are setup in the final pass, otherwise they are setup in this pass.
         bool m_HasFinalPass;
+		
+		// Some Android devices do not support sRGB backbuffer
+		// We need to do the conversion manually on those
+		bool m_EnableSRGBConversionIfNeeded;
 
         public PostProcessPass(RenderPassEvent evt, PostProcessData data)
         {
@@ -107,7 +111,7 @@ namespace UnityEngine.Rendering.Universal.Internal
             m_ResetHistory = true;
         }
 
-        public void Setup(in RenderTextureDescriptor baseDescriptor, in RenderTargetHandle source, in RenderTargetHandle destination, in RenderTargetHandle depth, in RenderTargetHandle internalLut, bool hasFinalPass)
+        public void Setup(in RenderTextureDescriptor baseDescriptor, in RenderTargetHandle source, in RenderTargetHandle destination, in RenderTargetHandle depth, in RenderTargetHandle internalLut, bool hasFinalPass, bool enableSRGBConversion)
         {
             m_Descriptor = baseDescriptor;
             m_Source = source;
@@ -116,14 +120,16 @@ namespace UnityEngine.Rendering.Universal.Internal
             m_InternalLut = internalLut;
             m_IsFinalPass = false;
             m_HasFinalPass = hasFinalPass;
+			m_EnableSRGBConversionIfNeeded = enableSRGBConversion;
         }
 
-        public void SetupFinalPass(in RenderTargetHandle source)
+        public void SetupFinalPass(in RenderTargetHandle source, bool enableSRGBConversion)
         {
             m_Source = source;
             m_Destination = RenderTargetHandle.CameraTarget;
             m_IsFinalPass = true;
             m_HasFinalPass = false;
+			m_EnableSRGBConversionIfNeeded = enableSRGBConversion;
         }
 
         public override void Configure(CommandBuffer cmd, RenderTextureDescriptor cameraTextureDescriptor)
@@ -311,7 +317,7 @@ namespace UnityEngine.Rendering.Universal.Internal
                 SetupGrain(cameraData.camera, m_Materials.uber);
                 SetupDithering(ref cameraData, m_Materials.uber);
 				
-				if (Display.main.requiresSrgbBlitToBackbuffer)
+				if (Display.main.requiresSrgbBlitToBackbuffer && m_EnableSRGBConversionIfNeeded)
 					m_Materials.uber.EnableKeyword(ShaderKeywordStrings.LinearToSRGBConversion);
 
                 // Done with Uber, blit it
@@ -954,7 +960,7 @@ namespace UnityEngine.Rendering.Universal.Internal
             SetupGrain(cameraData.camera, material);
             SetupDithering(ref cameraData, material);
 			
-			if (Display.main.requiresSrgbBlitToBackbuffer)
+			if (Display.main.requiresSrgbBlitToBackbuffer && m_EnableSRGBConversionIfNeeded)
 				material.EnableKeyword(ShaderKeywordStrings.LinearToSRGBConversion);
 
             cmd.SetGlobalTexture("_BlitTex", m_Source.Identifier());